requirements.txt
asgiref==3.5.0
Django==4.0.3
django-filter==21.1
python-dotenv==0.20.0
sqlparse==0.4.2
django-isbn-field==0.5.3